requirements.txt
flask
pandas
scikit-learn
nltk
joblib
google-generativeai